+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Marek Kurańda](https://github.com/mjkuranda): `InputList` component integrating `Autocomplete` from `MUI` library.
 - [Marek Kurańda](https://github.com/mjkuranda): `InputImage` component for uploading images.
 - [Marek Kurańda](https://github.com/mjkuranda): `InputSelect` component for selecting one of many options.
+- [Marek Kurańda](https://github.com/mjkuranda): `InputCheckbox` component for checking option.
 
 ## [1.30.0] - 2024-08-21
 ### Changed